requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
boto3==1.36.1
botocore==1.36.1
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
gcloud==0.18.3
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gotrue==2.11.1
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iso3166==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jws==0.1.3
mapbox==0.18.1
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
oauth2client==4.1.3
packaging==24.2
polyline==2.0.2
postgrest==0.19.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase3==4.0.2
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
realtime==2.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.11.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
websockets==13.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
